+++ ui/src/alerts/constants/tableSizing.js
@@ -1,7 +1,7 @@
 export const ALERTS_TABLE = {
-  colName: '15%',
-  colLevel: '10%',
-  colTime: '25%',
-  colHost: '25%',
-  colValue: '25%',
+  colName: 'calc(55% - 216px)',
+  colLevel: '62px',
+  colTime: '190px',
+  colHost: 'calc(45% - 196px)',
+  colValue: '160px',
 }