+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ### Features
 ### UI Improvements
+1. [#1707](https://github.com/influxdata/chronograf/pull/1707): Polish alerts table in status page to wrap text less
 
 ## v1.3.4.0 [2017-07-10]
 ### Bug Fixes